[break] merge BreakIE and ScreenJunkiesIE
